todolist/requirements.txt
absl-py==0.11.0
appdirs==1.4.4
asgiref==3.3.1
attrs==20.2.0
Babel==2.6.0
beautifulsoup4==4.9.2
cached-property==1.5.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
demjson==2.2.4
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.2
docutils==0.14
ebaysdk==2.1.5
feedparser==5.2.1
flatbuffers==1.12
gevent==1.4.0
google-pasta==0.2.0
greenlet==0.4.15
html2text==2018.1.9
isodate==0.6.0
Jinja2==2.10.1
jtutils==0.0.8
leven==1.0.4
libsass==0.17.0
lxml==4.5.2
Mako==1.0.7
MarkupSafe==1.1.0
mock==2.0.0
multitasking==0.0.9
nose==1.3.7
num2words==0.5.6
ofxparse==0.19
pandas==1.1.5
passlib==1.7.2
pbr==5.5.0
Pillow==6.2.1
polib==1.1.0
psutil==5.6.6
psycopg2==2.8.4
pycparser==2.20
pydot==1.4.1
PyMySQL==1.0.2
pyparsing==2.2.0
PyPDF2==1.26.0
pypiwin32==223
pyserial==3.4
python-csv==0.0.13
python-stdnum==1.14
pytz==2019.1
pyusb==1.0.2
qrcode==6.1
reportlab==3.5.13
requests-toolbelt==0.9.1
selenium==3.141.0
soupsieve==2.0.1
sqlparse==0.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
urllib3==1.26.4
vatnumber==1.2
vobject==0.9.6.1
xlrd==1.1.0
XlsxWriter==1.1.2
xlwt==1.3.0
xmltodict==0.12.0
yfinance==0.1.55
zeep==3.2.0